homecare {dunn.test}	R Documentation

Occupation and Home Care Eligibility

Description

Occupation and home care eligibility for 383 patients medically elgibile for home care.

Usage

homecare

Format

A data frame containing two variables, occupation and eligibility, over 383 observations.

Source

Dunn, O. J. (1964) Multiple comparisons using rank sums. Technometrics, 6(3), 241–252s. Table 1. <doi: 10.1080/00401706.1964.10490181>.

References

The data have been adapted from a study of a group of patients entering the Los Angeles County General Hospital during the years 1959-61.

[Package dunn.test version 1.3.7 Index]